+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Report errors finding vault from `:ObsidianCheckHealth`.
+
 ### Fixed
 
 - Removed annoying "skipped updating frontmatter" message on file write.